Corfu experiment partly failed for now
Seems to work in some modes but I cannot for the life of me get it
working in haskell-mode or even prog mode. I think because it uses the
native completion mechanism, it's subjected to all the fucked up, crufty
ways that different major modes interact with that system. I can kind of
get it working in emacs-lisp mode by forcing the completion-at-point
functions in those buffers but that does not work in haskell-mode.