Fix hang, use after! in some places
It seems like after! is generally preferable to use-package! for
packages that are provided by doom in init.el.

Bizarrely, if I deferred or commented out the synchronous sp-local-pair
in haskell_config.el, emacs would use all the available memory and crash
on boot. This somehow was due to global wrap, which I have turned off.