Add buffer switching to all modes
This was the simplest and least conflicting way I could figure out to
use the same keybinding to switch buffers across emacs, line mode, and
char mode. I'm pretty much only in char mode in terminals and maybe 100%
of the time when I switch it to line mode it's to switch buffers, or I
have to rotate to a window that isn't a terminal and switch buffers from
there. I'll have to adjust my muscle memory to use this new binding but
this seems worth it.