Rebind window splits
This makes logical sense and will allow me to do window splits that work
around windows in char mode without having to switch to line mode.
Switching to line mode to do window manipulations is particularly bad
because you'll lose focus of the window and when you return to it it
won't be in char mode like you'd want.